dummy_service/requirements.txt
asgiref==3.8.1
Django==5.0.9
djangorestframework==3.15.2
mssql-django==1.5
pyodbc==5.2.0
python-decouple==3.8
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
gunicorn==23.0.0